Change make_cuda_malloc to return a shared_ptr
I believe this was unique_ptr so that there isn't the burden of
the things that it didn't need. With unique_ptr this carries the
burden of the extra type signature because deleters aren't stored.
Change this, to make this easier to use, and if we notice an issue
we can cross that bridge.