Add raw_device_ptr passthrough and convert bitshuffle calls
Bitshuffle operates on a byte-level, so would be misleading to require an
explicit type (other than std::byte). But the shared_ptr only points to
a specific pointer type, and we don't want to expose getting the value of
the pointer to regular C++. So, have a raw pointer that converts automatically.

Man, I'd love Borrow/AsRef/Deref.