Fix pointer arithmetic bug with LZ4 compression
Before the smart-pointers refactor, this was using an array of uint16_t
to store the compressed bitshuffle data. With the refactor, this was changed
to an array of bytes because that better represents the encoded state of
the bitshuffled buffer. This, however, meant that the pointer arithmetic was
operating in bytes rather than uint16 now, and so the numbers were wrong.